chore: move responseHandler to own file
